[PORT FROM MAIN] Music: Modify the DownloadProvider to support the drm same file problem.
BZ: 52128

R3-Change-Id: I63cf3f4d594fd1a3a4d9fb99804c73e630bc3481
Borqs-Id: 032e6c2121796e03be2e509f20e1cdfcbcd625fe
Borqs-Contact: b618 <b618@borqs.com>
Date: Fri, 8 Jun 2012 11:33:44 +0800
rootcause: When we download the a drm file such as 1.dm.
we'll verify whether it is already exist and then modify the extension to fl such as 1.fl.
When we download again 1.dm, and verify if it's exist. But it has already been modified to 1.fl.
We should modify the extension first and then getfullpath of the file.
Reviewed-on: https://bkbgerrit.borqs.com/gerrit/4128
Tested-by: svnadmin <svnadmin@borqs.com>
Reviewed-by: b052 <b052@borqs.com>
Reviewed-by: b169 <b169@borqs.com>

Change-Id: I131e38074f79f60e5c738c62da9a589685314184
Reviewed-on: http://android.intel.com:8080/76007
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>