DownloadNotifier: Fix cannot view detail of downing files under some conditions.
BZ: 117417

When download a same file more than once, view detail of downing files
from the StatusBar pull-down menu will failed.

Root cause of this issue is when DownloadNotifier send a nitification to
NotificationManager, it use the dafault flags values to get broadcast
from PendingIntent. With that value, the extra datas of broadcast intent
will not get updated. Finally, the it failed to open download files.

Set flags to PendingIntent.FLAG_UPDATE_CURRENT when get broadcast from
PendingIntent to fix this issue.

Change-Id: I63fa34d7742156f888e51ba19626123e433f034e
Signed-off-by: lwan89x <liangx.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/117397
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>